refactored code, save, close, and edit indv stickies
